BIG refactoring
:repeat: Edit : a lot of things ;
:white_check_mark: Create : a lot of files.
