+++ commands/server.js
@@ -45,7 +45,8 @@ module.exports = {
     args: false,
     usage: 'o*server',
     guildOnly: true,
-    adminOnly: false,
+    moderatorOnly: false,
+    creatorOnly: false,
     cooldown: 60,
     execute(message, args) {
         const dateRaw = message.guild.createdAt;
